refactor: trim cron session store startup imports
